code improvements
