+++ app/views/npq_separation/admin/finance/statements/show.html.erb
@@ -40,7 +40,7 @@
           @statement.ecf_id
         end %>
 
-    <% if @statement.has_marked_as_paid_at_date? %>
+    <% if @statement.marked_as_paid_with_date? %>
       <p class="govuk-!-margin-bottom-2">
         <strong class="govuk-tag govuk-tag--green">Authorised for payment at <%= @statement.marked_as_paid_at.in_time_zone("London").strftime("%-I:%M%P on %-e %b %Y") %></strong>
       </p>
